refactor(option-A): Phase 4~6 — UserRecipe rename + Long→Integer 전역 ripple
Phase 4 (UserRecipe rename): PendingRecipe 도메인 통째 UserRecipe 로 git mv
+ 클래스/테이블/PK 컬럼/URL/필드명 일괄 sed. 영향: entity/repository/service/
controller + 6개 DTO (UserRecipe/AdminUserRecipe*). URL /api/v1/pending-recipes
→ /api/v1/user-recipes, /admin/pending-recipes → /admin/user-recipes.

Phase 5 (PK Long → Integer): Recipe/RecipeIngredient/Step/Label/Media/Stats,
Like/Scrap, ChatRoom/ChatMessage entity 들의 PK·FK Integer 화 + repository
JpaRepository<X,Long> → <X,Integer> + 메서드 시그니처 + Collection<Long>/
Set<Long>/Map<Long,X>/Optional<Long> generic 인자 일괄 Integer 화.

Phase 6 (cross-cutting): UserMeService 의 socialAccountRepository →
userIdentityRepository, user.getDeletedAt() != null → !user.isActive()
(탈퇴 표식 isActive 단일화), AdminUserService 동일 정정. Controllers
@PathVariable Long → Integer, currentUserId() helper Integer.

별도 처리: KakaoOAuthClient.KakaoApiResponse.id 는 카카오 측 64bit 라
Long 유지 (overflow 방지) — providerUserId 로 String 변환 시점에 안전.

RecipeService/ScrapService cursor parse Long.parseLong → Integer.parseInt
+ 'likes' 커서 long[] → int[] 로 정합. main compileJava PASS.

+++ src/main/java/com/naengo/api_server/domain/admin/controller/AdminUserController.java
@@ -14,13 +14,13 @@ public class AdminUserController {
 
     /** SPEC-20260504-03 — 사용자 차단. */
     @PostMapping("/{userId}/block")
-    public AdminUserBlockResponse block(@PathVariable Long userId) {
+    public AdminUserBlockResponse block(@PathVariable Integer userId) {
         return adminUserService.block(userId);
     }
 
     /** SPEC-20260504-03 — 차단 해제. */
     @PostMapping("/{userId}/unblock")
-    public AdminUserBlockResponse unblock(@PathVariable Long userId) {
+    public AdminUserBlockResponse unblock(@PathVariable Integer userId) {
         return adminUserService.unblock(userId);
     }
 }
