+++ src/main/java/com/naengo/api_server/domain/admin/dto/AdminUserBlockResponse.java
@@ -1,4 +1,4 @@
 package com.naengo.api_server.domain.admin.dto;
 
-public record AdminUserBlockResponse(Long userId, boolean isBlocked) {
+public record AdminUserBlockResponse(Integer userId, boolean isBlocked) {
 }
