+++ src/main/java/com/naengo/api_server/domain/recipe/controller/RecipeController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.*;
 /**
  * 공개 레시피 조회 — api-3.json `/api/v1/recipes`.
  *
- * <p>사용자 제출(pending) 흐름은 {@link PendingRecipeController} 책임.
+ * <p>사용자 제출(pending) 흐름은 {@link UserRecipeController} 책임.
  */
 @RestController
 @RequestMapping("/api/v1/recipes")
@@ -29,7 +29,7 @@ public class RecipeController {
 
     /** 단건 조회 (recipes 만). 응답은 목록 항목과 동형 (api-3.json 정합). */
     @GetMapping("/{id}")
-    public RecipeListItemResponse detail(@PathVariable Long id) {
+    public RecipeListItemResponse detail(@PathVariable Integer id) {
         return recipeService.detail(id);
     }
 }
