+++ src/main/java/com/naengo/api_server/domain/recipe/repository/RecipeStatsRepository.java
@@ -3,5 +3,5 @@ package com.naengo.api_server.domain.recipe.repository;
 import com.naengo.api_server.domain.recipe.entity.RecipeStats;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecipeStatsRepository extends JpaRepository<RecipeStats, Long> {
+public interface RecipeStatsRepository extends JpaRepository<RecipeStats, Integer> {
 }
